requirements.txt
pymupdf==1.23.8
kss==4.5.0
pandas==2.1.4
openpyxl==3.1.2
rapidfuzz==3.6.1
difflib-json==0.1.0
pyyaml==6.0.1
pytesseract==0.3.10
pillow==10.1.0
git+https://github.com/ssut/py-hanspell.git
pykospacing==0.5.0
